cast types in data generation

+++ data_generation/main.py
@@ -19,8 +19,8 @@ def generate_dataset(dataset, num_samples, path, params):
 
     for i in tqdm(range(num_samples)):
         input, solution = generate_sample(i, **params)
-        np.save(os.path.join(path, dataset, f"input_{str(i).rjust(l, '0')}"), input)
-        np.save(os.path.join(path, dataset, f"solution_{str(i).rjust(l, '0')}"), solution)
+        np.save(os.path.join(path, dataset, f"input_{str(i).rjust(l, '0')}"), input.astype(np.float32))
+        np.save(os.path.join(path, dataset, f"solution_{str(i).rjust(l, '0')}"), solution.astype(np.float32))
 
 
 def main():
